fix(conv_thread_impl.rs and conv_data_impl.rs): fix shared data

+++ compiler/AADLSource/pingpong/ping.c
@@ -3,14 +3,14 @@
 
 int p = 0;
 
-void user_do_ping_spg(long long *v) {
+void user_do_ping_spg(int *v) {
     printf("*** SENDING PING *** %d\n", p);
     *v = p;
     p++;
     fflush(stdout);
 }
 
-void user_ping_spg(long long i) {
+void user_ping_spg(int i) {
     printf("*** PING *** %d\n", i);
     fflush(stdout);
 }
